NBL-24 Fix conversions between long and int64_t

+++ modules/platforms/cpp/core/src/impl/cluster/cluster_group_impl.cpp
@@ -570,7 +570,7 @@ namespace ignite
             {
                 RefreshNodes();
 
-                return static_cast<long>(topVer);
+                return topVer;
             }
 
             SP_ClusterGroupImpl ClusterGroupImpl::GetEmptyClusterGroup()
